\name{length.gene.hg19}
\alias{length.gene.hg19}
\docType{data}
\title{
Collapsed Gene Length from UCSC HG19 annotation
}
\description{
Collapsed Gene Length from UCSC HG19 annotation
}
\usage{data("length.gene.hg19")}
\format{
  The format is:
 int [1:47352, 1] 230 320 1774 2134 9544 2907 4689 2190 5400 454 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:47352] "5S_rRNA" "7SK" "A1BG" "A1BG-AS1" ...
  ..$ : chr "Length"
}

\keyword{datasets}
