\name{length.isoform.hg19}
\alias{length.isoform.hg19}
\docType{data}
\title{
Transcript Length from UCSC HG19 annotation
}
\description{
Transcript Length from UCSC HG19 annotation
}
\usage{data("length.isoform.hg19")}
\format{
  The format is:
 int [1:82960, 1] 1652 1746 2412 1769 1577 1130 918 32 62 4370 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:82960] "uc001aaa.3" "uc001aac.4" "uc001aae.4" "uc001aah.4" ...
  ..$ : chr "Length"
}

\keyword{datasets}
