\name{x}
\alias{x}
\docType{data}
\title{
Counts Matrix
}
\description{
Counts Matrix
}
\usage{data("x")}
\format{
  A data frame with 15998 observations on the following 14 variables.
  \describe{
    \item{\code{X127_SRR037452}}{a numeric vector}
    \item{\code{X127_SRR037453}}{a numeric vector}
    \item{\code{X127_SRR037454}}{a numeric vector}
    \item{\code{X127_SRR037455}}{a numeric vector}
    \item{\code{X127_SRR037456}}{a numeric vector}
    \item{\code{X127_SRR037457}}{a numeric vector}
    \item{\code{X127_SRR037458}}{a numeric vector}
    \item{\code{X128_SRR037459}}{a numeric vector}
    \item{\code{X128_SRR037460}}{a numeric vector}
    \item{\code{X128_SRR037461}}{a numeric vector}
    \item{\code{X128_SRR037462}}{a numeric vector}
    \item{\code{X128_SRR037463}}{a numeric vector}
    \item{\code{X128_SRR037464}}{a numeric vector}
    \item{\code{X128_SRR037465}}{a numeric vector}
  }
}

\keyword{datasets}
